# backend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "object"
              }
            ]
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "EventCreate": {
        "example": {
          "delta": 3,
          "event_type": "SENSOR_IN",
          "idempotency_key": "evt-001",
          "location": "ALM-CENTRAL",
          "product_id": 1,
          "source": "sensor_simulado"
        },
        "properties": {
          "delta": {
            "examples": [
              3
            ],
            "exclusiveMinimum": 0.0,
            "title": "Delta",
            "type": "integer"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventType",
            "examples": [
              "SENSOR_IN"
            ]
          },
          "idempotency_key": {
            "examples": [
              "evt-001"
            ],
            "maxLength": 100,
            "minLength": 1,
            "title": "Idempotency Key",
            "type": "string"
          },
          "location": {
            "default": "default",
            "examples": [
              "ALM-CENTRAL"
            ],
            "maxLength": 100,
            "minLength": 1,
            "title": "Location",
            "type": "string"
          },
          "product_id": {
            "examples": [
              1
            ],
            "title": "Product Id",
            "type": "integer"
          },
          "source": {
            "default": "sensor_simulado",
            "examples": [
              "sensor_simulado"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "event_type",
          "product_id",
          "delta",
          "idempotency_key"
        ],
        "title": "EventCreate",
        "type": "object"
      },
      "EventListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "EventListResponse",
        "type": "object"
      },
      "EventResponse": {
        "example": {
          "created_at": "2026-02-17T10:30:00Z",
          "delta": 3,
          "event_status": "PENDING",
          "event_type": "SENSOR_IN",
          "id": 55,
          "processed": false,
          "product_id": 1,
          "source": "sensor_simulado"
        },
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "delta": {
            "title": "Delta",
            "type": "integer"
          },
          "event_status": {
            "title": "Event Status",
            "type": "string"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventType"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "processed": {
            "title": "Processed",
            "type": "boolean"
          },
          "product_id": {
            "title": "Product Id",
            "type": "integer"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "id",
          "event_type",
          "product_id",
          "delta",
          "source",
          "processed",
          "event_status",
          "created_at"
        ],
        "title": "EventResponse",
        "type": "object"
      },
      "EventType": {
        "enum": [
          "SENSOR_IN",
          "SENSOR_OUT"
        ],
        "title": "EventType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MovementOperation": {
        "example": {
          "location": "ALM-CENTRAL",
          "movement_source": "SCAN",
          "product_id": 1,
          "quantity": 5
        },
        "properties": {
          "location": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Location",
            "type": "string"
          },
          "movement_source": {
            "$ref": "#/components/schemas/Source"
          },
          "product_id": {
            "title": "Product Id",
            "type": "integer"
          },
          "quantity": {
            "exclusiveMinimum": 0.0,
            "title": "Quantity",
            "type": "integer"
          }
        },
        "required": [
          "product_id",
          "quantity",
          "location",
          "movement_source"
        ],
        "title": "MovementOperation",
        "type": "object"
      },
      "MovementResponse": {
        "example": {
          "created_at": "2026-02-17T10:20:00Z",
          "delta": 5,
          "id": 100,
          "location": "ALM-CENTRAL",
          "location_id": 1,
          "movement_source": "SCAN",
          "movement_type": "IN",
          "product_id": 1,
          "quantity": 5,
          "user_id": 2
        },
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "delta": {
            "title": "Delta",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "location_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "movement_source": {
            "$ref": "#/components/schemas/Source"
          },
          "movement_type": {
            "$ref": "#/components/schemas/MovementType"
          },
          "product_id": {
            "title": "Product Id",
            "type": "integer"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "transfer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transfer Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "id",
          "product_id",
          "quantity",
          "delta",
          "movement_type",
          "movement_source",
          "user_id",
          "location_id",
          "location",
          "created_at"
        ],
        "title": "MovementResponse",
        "type": "object"
      },
...
      "MovementType": {
        "enum": [
          "IN",
          "OUT",
          "ADJUST"
        ],
        "title": "MovementType",
        "type": "string"
      },
      "MovementWithStockResponse": {
        "properties": {
          "movement": {
            "$ref": "#/components/schemas/MovementResponse"
          },
          "stock": {
            "$ref": "#/components/schemas/StockResponse"
          }
        },
        "required": [
          "stock",
          "movement"
        ],
        "title": "MovementWithStockResponse",
        "type": "object"
      },
...
      "Source": {
        "enum": [
          "SCAN",
          "MANUAL"
        ],
        "title": "Source",
        "type": "string"
      },
...
      "StockResponse": {
        "example": {
          "created_at": "2026-02-17T10:00:00Z",
          "id": 7,
          "location": "ALM-CENTRAL",
          "product_id": 1,
          "quantity": 10,
          "updated_at": "2026-02-17T10:10:00Z"
        },
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "location": {
            "examples": [
              "ALM-CENTRAL"
            ],
            "maxLength": 100,
            "minLength": 1,
            "title": "Location",
            "type": "string"
          },
          "product_id": {
            "examples": [
              1
            ],
            "title": "Product Id",
            "type": "integer"
          },
          "quantity": {
            "examples": [
              10
            ],
            "minimum": 0.0,
            "title": "Quantity",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "product_id",
          "location",
          "quantity",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "StockResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Sistema Inventariado Sensores",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/events/": {
      "get": {
        "operationId": "list_events_events__get",
        "parameters": [
          {
            "in": "query",
            "name": "event_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EventType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Type"
            }
          },
          {
            "in": "query",
            "name": "product_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Id"
            }
          },
          {
            "in": "query",
            "name": "processed",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Processed"
            }
          },
          {
            "in": "query",
            "name": "order_by",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "created_at",
              "title": "Order By"
            }
          },
          {
            "in": "query",
            "name": "order_dir",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "desc",
              "title": "Order Dir"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "order_by debe ser 'created_at'"
                }
              }
            },
            "description": "Ordenacion invalida"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Events",
        "tags": [
          "events"
        ]
      },
      "post": {
        "operationId": "create_event_events__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "created_at": "2026-02-17T10:30:00Z",
                  "delta": 3,
                  "event_status": "PENDING",
                  "event_type": "SENSOR_IN",
                  "id": 55,
                  "processed": false,
                  "product_id": 1,
                  "source": "sensor_simulado"
                }
              }
            },
            "description": "Evento ya existente (idempotencia)"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Producto no encontrado"
                }
              }
            },
            "description": "Producto no encontrado"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "location es obligatorio"
                }
              }
            },
            "description": "Payload invalido"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "No se pudo encolar el evento: broker no disponible"
                }
              }
            },
            "description": "No se pudo encolar el evento"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Event",
        "tags": [
          "events"
        ]
      }
    },
...
    "/movements/out": {
      "post": {
        "operationId": "movement_out_movements_out_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovementOperation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovementWithStockResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Stock insuficiente"
                }
              }
            },
            "description": "Regla de negocio invalida"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Movement Out",
        "tags": [
          "movements"
        ]
      }
    },
...
  }
}